requirements.txt
numpy
matplotlib
paho-mqtt
adafruit-circuitpython-dht
board